requirements.txt
kafka-python==2.1.0
python-dateutil
